libs/superagent/pyproject.toml
[tool.poetry]
name = "superagent"
version = "0.2.14"
description = "🥷 Run AI-agents with an API"
authors = ["Ismail Pelaseyed"]
readme = "../../README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.8.1, <3.12"
fastapi = { extras = ["all"], version = "^0.95.1" }
uvicorn = { extras = ["standard"], version = "^0.22.0" }
gunicorn = "^20.1.0"
python-decouple = "^3.8"
pydantic = "^1.10.7"
sse-starlette = "^1.5.0"
bcrypt = "^4.0.1"
pyjwt = "^2.7.0"
ruff = "^0.0.265"
black = "^23.3.0"
uuid = "^1.30"
flake8 = "^6.0.0"
tiktoken = "0.5.2"
python-slugify = "^8.0.1"
asyncio = "^3.4.3"
colorlog = "^6.7.0"
vulture = "^2.7"
lamini = "^0.0.21"
pytest = "^7.4.0"
prefect = "^2.11.4"
llama-index = "^0.8.11.post3"
pinecone-client = "^2.2.2"
pyairtable = "^2.1.0.post1"
backoff = "^2.2.1"
xmltodict = "^0.13.0"
pypdf = "^3.16.0"
tabulate = "^0.9.0"
gitpython = "^3.1.36"
python-pptx = "^0.6.22"
openpyxl = "^3.1.2"
unstructured = "^0.10.16"
requests = "^2.31.0"
wolframalpha = "^5.0.0"
bs4 = "^0.0.1"
segment-analytics-python = "^2.2.3"
replicate = "^0.15.4"
e2b = "^0.10.8"
youtube-transcript-api = "^0.6.1"
openai = "^1.1.1"
langchain-experimental = "^0.0.37"
pydub = "^0.25.1"
algoliasearch = "^3.0.0"
litellm = "^1.14.1"
weaviate-client = "^3.25.3"
qdrant-client = "^1.6.9"
vecs = "^0.4.2"
oauth2client = "^4.1.3"
agentops = {extras = ["langchain"], version = "^0.0.20"}
langchain = "^0.1.4"
langchain-openai = "^0.0.5"
python-docx = "^1.1.0"
prisma = "^0.12.0"
stripe = "^8.2.0"
redis = "^5.0.1"
langsmith = "^0.1.9"
langfuse = "^2.16.2"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.vulture]
exclude = [
    "*settings.py",
    "*/docs/*.py",
    "*/test_*.py",
    "*/.venv/*.py",
    "app/lib/callbacks.py", # Ignoring this file
]
ignore_decorators = ["@app.route", "@require_*"]
ignore_names = ["visit_*", "do_*"]
make_whitelist = true
min_confidence = 100
paths = ["."]
sort_by_size = true
verbose = false

[tool.ruff]
exclude = [
    "*settings.py",
    "*/docs/*.py",
    "*/test_*.py",
    "*/.venv/*.py",
    "whitelist.py",
]
